Added PhreakNodeOperations
